Add files via upload
The talker counts 10 times a second starting from 1 and the listener confirms this number. When talker is at 69, listener says "nice" 31 times. Even thou I managed this awesome joke, I could not implement a quit() command into listener. I reckon this comment only halts a certain other function when it is in the if statement, but the details should be discussed in the lab.